+++ VSIX/RapidXaml.Analysis/XamlAnalysis/CustomAnalysis/Issue364ExampleAnalyzer.cs
@@ -1,13 +1,17 @@
 ﻿// Copyright (c) Matt Lacey Ltd. All rights reserved.
 // Licensed under the MIT license.
 
-using System.Collections.Generic;
 using RapidXaml;
 
 namespace RapidXamlToolkit.XamlAnalysis.CustomAnalysis
 {
     public class Issue364ExampleAnalyzer : BuiltInXamlAnalyzer
     {
+        public Issue364ExampleAnalyzer(VisualStudioIntegration.IVisualStudioAbstraction vsa)
+            : base(vsa)
+        {
+        }
+
         public override string TargetType() => "ANYCONTAINING:=\"{Binding";
 
         public override AnalysisActions Analyze(RapidXamlElement element, ExtraAnalysisDetails extraDetails)