+++ VSIX/RapidXaml.Analysis/XamlAnalysis/CustomAnalysis/XamarinForms/LabelAnalyzer.cs
@@ -1,37 +1,41 @@
 ﻿// Copyright (c) Matt Lacey Ltd. All rights reserved.
 // Licensed under the MIT license.
 
-using System;
-using System.Linq;
 using RapidXaml;
-using RapidXamlToolkit.Resources;
 
 namespace RapidXamlToolkit.XamlAnalysis.CustomAnalysis
 {
     public class LabelAnalyzer : BuiltInXamlAnalyzer
     {
+        public LabelAnalyzer(VisualStudioIntegration.IVisualStudioAbstraction vsa)
+            : base(vsa)
+        {
+        }
+
         public override string TargetType() => Elements.Label;
 
         public override AnalysisActions Analyze(RapidXamlElement element, ExtraAnalysisDetails extraDetails)
         {
-            var txtAttr = element.GetAttributes(Attributes.Text).FirstOrDefault();
+            return this.CheckForHardCodedString(Attributes.Text, AttributeType.Any, element, extraDetails);
 
-            if (txtAttr != null && txtAttr.HasStringValue)
-            {
-                var value = txtAttr.StringValue;
+            ////var txtAttr = element.GetAttributes(Attributes.Text).FirstOrDefault();
 
-                // TODO: ISSUE#163 change this to an RXT200 when can handle localization of Xamarin.Forms apps
-                if (!string.IsNullOrWhiteSpace(value) && char.IsLetterOrDigit(value[0]))
-                {
-                    return AnalysisActions.HighlightAttributeWithoutAction(
-                    errorType: RapidXamlErrorType.Warning,
-                    code: "RXT201",
-                    description: StringRes.UI_XamlAnalysisGenericHardCodedStringDescription.WithParams(Elements.Label, Attributes.Text, value),
-                    attribute: txtAttr);
-                }
-            }
+            ////if (txtAttr != null && txtAttr.HasStringValue)
+            ////{
+            ////    var value = txtAttr.StringValue;
 
-            return AnalysisActions.None;
+            ////    // TODO: ISSUE#163 change this to an RXT200 when can handle localization of Xamarin.Forms apps
+            ////    if (!string.IsNullOrWhiteSpace(value) && char.IsLetterOrDigit(value[0]))
+            ////    {
+            ////        return AnalysisActions.HighlightAttributeWithoutAction(
+            ////        errorType: RapidXamlErrorType.Warning,
+            ////        code: "RXT201",
+            ////        description: StringRes.UI_XamlAnalysisGenericHardCodedStringDescription.WithParams(Elements.Label, Attributes.Text, value),
+            ////        attribute: txtAttr);
+            ////    }
+            ////}
+
+            ////return AnalysisActions.None;
         }
     }
 }