+++ VSIX/RapidXaml.CustomAnalysis/ActionType.cs
@@ -44,5 +44,15 @@ namespace RapidXaml
         /// Replace the entire element.
         /// </summary>
         ReplaceElement,
+
+        /// <summary>
+        /// Create an entry in a Resource file (RESW or RESX).
+        /// </summary>
+        CreateResource,
+
+        /// <summary>
+        /// Remove the body content of the element.
+        /// </summary>
+        RemoveContent,
     }
 }