+++ VSIX/RapidXaml.Shared/VisualStudioIntegration/IVisualStudioAbstraction.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Matt Lacey Ltd. All rights reserved.
 // Licensed under the MIT license.
 
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace RapidXamlToolkit.VisualStudioIntegration
@@ -18,5 +19,9 @@ namespace RapidXamlToolkit.VisualStudioIntegration
         (int, int) GetCursorPositionAndLineNumber();
 
         Task<int> GetXamlIndentAsync();
+
+        string GetLanguageFromContainingProject(string fileName);
+
+        List<string> GetFilesFromContainingProject(string fileName, params string[] fileNameEndings);
     }
 }