+++ VSIX/RapidXamlToolkit.Tests/XamlAnalysis/CustomAnalyzers/XamarinForms/LabelAnalyzerTests.cs
@@ -17,11 +17,13 @@ namespace RapidXamlToolkit.Tests.XamlAnalysis.CustomAnalyzers
 
             var actions = this.GetActions<LabelAnalyzer>(xaml, ProjectType.XamarinForms);
 
-            // TODO: ISSUE#163 update when add support for localizing hard-coded strings in Xamarin.Forms.
             Assert.IsFalse(actions.IsNone);
             Assert.AreEqual(1, actions.Actions.Count);
-            Assert.AreEqual(ActionType.HighlightWithoutAction, actions.Actions[0].Action);
-            Assert.AreEqual("RXT201", actions.Actions[0].Code);
+            Assert.AreEqual(ActionType.CreateResource, actions.Actions[0].Action);
+            Assert.AreEqual("RXT200", actions.Actions[0].Code);
+            Assert.AreEqual(2, actions.Actions[0].SupplementaryActions.Count);
+            Assert.AreEqual(ActionType.RemoveAttribute, actions.Actions[0].SupplementaryActions[0].Action);
+            Assert.AreEqual(ActionType.AddAttribute, actions.Actions[0].SupplementaryActions[1].Action);
         }
     }
 }