+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix bug where the location of items to underline was not calculated correctly.
 - Performance improvements in analysis.
+- Add JsonSchema support for .xamlAnalysis files
 
 ## 0.11.4